refac: extracted common code into hook
